+++ Tests/Mock/MockPlayerItem.swift
@@ -28,27 +28,44 @@ import AVFoundation
 
 final class MockPlayerItem: AVPlayerItem {
 
+    convenience init(url: URL, duration: CMTime?, status: AVPlayerItemStatus?) {
+        self.init(url: url)
+        overrideDuration = duration
+        overrideStatus = status
+    }
+    
+    convenience init(asset: AVAsset, duration: CMTime?, status: AVPlayerItemStatus?) {
+        self.init(asset: asset)
+        overrideDuration = duration
+        overrideStatus = status
+    }
+    
     var overrideStatus: AVPlayerItemStatus?
     override var status: AVPlayerItemStatus {
-        return overrideStatus ?? AVPlayerItemStatus.unknown
+        return overrideStatus ?? .unknown
     }
     
     private(set) var cancelPendingSeeksCallCount = 0
     override func cancelPendingSeeks() {
         cancelPendingSeeksCallCount += 1
     }
+    
+    private(set) var overrideDuration: CMTime?
+    override var duration: CMTime {
+        return overrideDuration ?? kCMTimeZero
+    }
 }
 
 extension MockPlayerItem {
     
-    static func createOne(url: String) -> MockPlayerItem {
+    static func createOne(url: String, duration: CMTime? = nil, status: AVPlayerItemStatus? = nil) -> MockPlayerItem {
         let url = URL(string: url)!
-        return MockPlayerItem(url: url)
+        return MockPlayerItem(url: url, duration: duration, status: status)
     }
     
-    static func createOnUsingAsset(url: String) -> MockPlayerItem {
+    static func createOnUsingAsset(url: String, duration: CMTime? = nil, status: AVPlayerItemStatus? = nil) -> MockPlayerItem {
         let url = URL(string: url)!
         let asset = MockAVAsset(url: url)
-        return MockPlayerItem(asset: asset)
+        return MockPlayerItem(asset: asset, duration: duration, status: status)
     }
 }